asterisk-11.x-chan-dongle: fix audio endianess problem
Github user @ljakob supplied a patch in
https://github.com/openwrt/telephony/issues/7
--
Hi, I've solved the problem myself - it was an endianness bug:

Here what gave me the correct clue:

    audio is broken in one direction only
    audio is fine if you disable any timing-source within asterisk (no good idea if you want to use IAX)
    audio was fine on my x86 box
    I had a long look into the code :)

The bug is in the audio handling of timing_write in channel.c, the data sent to the socket is not changed to the correct endianness if a timing device is used.

Here is my patch, it's untested (just one call yet) but it's trivial enough:
http://www.hugo.weite-welt.com/asterisk13_chan_dongle_endianess.patch
it is against asterisk13 branch in
https://github.com/oleg-krv/asterisk-chan-dongle.git
but should fit into other asterisk versions nicely
--
This commit imports the patch above to asterisk-11.x-chan-dongle.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>