yate-scripts-perl: Make perlbase-data optional
Package yate-scripts-perl requires package perlbase-data to work
properly, but perlbase-data is only needed for function dump() which I
expect to be used only when debugging YATE modules. It's possible to
write YATE modules without using the dump() function and then it's a
waste of resources to require the user to have perlbase-data installed
since it will consume memory while the YATE module is running and disk
space for the package installation.

So, let's change Yate.pm to only load Data::Dumper (from perlbase-data)
when the dump() function is called. This way Yate.pm can be used even
if perlbase-data isn't installed, but if you want to use the dump()
function you of course need to install perlbase-data.

Signed-off-by: Robert Högberg <robert.hogberg@gmail.com>