net/asterisk-11.x and net/asterisk-13.x: remove $(SDK) from ifneqs
The variable prevents the ifneqs from working on the buildbots. $(SDK)
is set there, so the ifneqs that test for empty will always be true.

libpq for instance doesn't build on aarch64. So the asterisk pgsql
package has a dependency on @!aarch64. An ifneq is used to disable libpq
detection by the configure script, to prevent the build to fail. But
because of $(SDK) being set the ifneq is always true, so libpq detection
is enabled, causing the build to always fail.

There are only two other packages in the tree that actually use the
$(SDK) variable, so it's fair to assume it's not needed for the asterisk
packages and can just be dropped.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>