net/asterisk-11.x and net/asterisk-13.x: fix menuselect
menuselect often fails on the buildbots:

menuselect/menuselect: error while loading shared libraries: libxml2.so.2: cannot open shared object file: No such file or directory

This doesn't happen on all builds. The assumption is that on the
particular buildbot where there is no error the buildbots own libxml2.so
is usable (it exists and the version is suitable).

To fix this make the linker add an rpath to the menuselect utility,
pointing to the hostpkg lib dir.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>