net/asterisk-11.x and net/asterisk-13.x: fix LOW_MEMORY, drop patch
LOW_MEMORY can't be enabled by defining it in ASTCFLAGS. This can be
verified in the headers Asterisk installs: LOW_MEMORY actually is not
defined (it would be if LOW_MEMORY was enabled).

LOW_MEMORY can be enabled successfully with menuselect. This commit
makes menuselect usable and calls it to enable LOW_MEMORY.

This commit also disables BUILD_NATIVE with menuselect. So even if
configure detects that -march=native is available, it won't be used.
This means the configure.ac patch that was added previously can be
dropped.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>