asterisk11: fix conffiles define
The configuration file protection for modules does not work when there
is more than one configuration file. Currently conffiles is populated
with a space separated list. But the items need to be separated by
newlines. With spaces, only the first item gets added to conffiles. The
remaining items are dropped and not protected. Fix this by replacing
spaces with newlines.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>