asterisk-chan-lantiq: Move to separate package
Currently chan-lantiq doesn't build on the build bots and users end up
having to build it themselves. The reason for this was explained by
Mathias Kresin here:
http://lists.infradead.org/pipermail/lede-dev/2017-September/008886.html.

To fix this put chan-lantiq into its own package and set
PKG_FLAGS:=nonshared. The new package uses variants already, so it is
easily extendable in the future.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>