chan-lantiq: setup Lantiq kernel headers locally
This commit unpacks the sources of lib_ifxos, drv_tapi and drv_vmmc,
patches them and sets up their header files for the compiler to use when
building chan-lantiq.

The dependency on kmod-ltq-vmmc is dropped so that the package can be
built by the buildbots. Instead, a dependency on mips is added so that
the buildbots will only build the package for the ARCH that Lantiq
actually runs on.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>