asterisk-15.x: fix CONFIGURE_ARGS
The way CONFIGURE_ARGS are done for srtp and pjproect does not take into
account that more than one package depends on them. This can result in
build failures. So add ifeqs to test properly if --with or --without is
required.

This commit also adds CONFIGURE_ARGS checks for:

alsa (asound)
portaudio
speex
speexdsp

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>