asterisk-13.x: speed up menuselect calls
Currently the menuselect calls, for each category or item, step into a
directory and call menuselect once. So if all packages are build
menuselect will be called hundreds of times and as many directory
changes will be done.

Instead step into the directory only once, then generate the argument
string by iteratively adding to it and then call menuselect once. The
speedup is big.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>