asterisk-13.x: workaround for build failure
On the buildbots (and locally, too) the x86 builds currently fail. It's
visible that the Asterisk build system attempts to build menuselect a
second time, which fails because it follows the standard recipes in its
Makefiles. The first build of menuselect on the other hand only works
because the OpenWrt package sets up the build properly.

The second build attempt happens because the menuselect CFLAGS category
is setup to delete a stamp file when anything in it changes. On x86
-march=native is available, so the OpenWrt package disables that via
menuselect, hence this item in the category is changed and the stamp
file automatically removed. The missing stamp file triggers a rebuild of
menuselect (which is not required for OpenWrt's intents and purposes).
Another item in the CFLAGS category with which the package can have the
same issue is LOW_MEMORY, but that is not enabled (changed) by default.

Work around this problem by recreating the stamp file after the
menuselect calls.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>