freeswitch: adjust python mk files
Since commit:
  https://github.com/openwrt/packages/commit/ccdc6bc530444283a510ceb6a50878493fe356f2
the python package.mk files have moved.

The intent is to also get rid of the include_mk
construct, because that one relies for python
to be built, before having access to those mk files.

There might be another change to the approach
of using python package mk files, but until
then, they need to be included directly.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>