kamailio-5.x: improve module selection
Currently INCL_MODULES is filled with all modules. This commit changes
this so that only the modules that are selected are added to it.

To further limit the number of modules built group_include="standard" is
added. This way only the modules of the standard group plus the modules
selected in menuconfig are built. The "standard" group only includes
modules without any depends.

This fixes build failures on targets where not all depends are
available, i.e. postgresql on ARC. Now these modules can be properly
prevented from being built.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>