freeswitch-stable: workaround for build failures when gcc 7 is used
When gcc 7 is used, -Werror=format-truncation finds snprintf() calls in
mod_cdr_mongodb.c and suspects that they may result in truncation. As a
result the build fails.

Work around this by adding -Wno-error=format-truncation to CFLAGS when
gcc support this.

JIRA-10939 has been raised upstream to follow up on this.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>