yate: turn miniwebrtc error into warning
ilbcwebrtc and isaccodec currently depend on @!arc because the code does
not compile there. The code does not compile on aarch64 either.

Instead of adding @!aarch64 this commit adds a patch copied from Debian
that turns the error "Please add support for your architecture in
typedefs.h" into a warning, allowing the code to build.

Consequently this commit removes @!arc from the depends as well.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>