kamailio-5.x: make internal libs available always
kamailio comes with internal libraries. Some are not built when only the
standard group of modules is compiled (i.e. no modules selected in
menuconfig).

The problem is that there is no way to rely on specific internal
libraries being available in the main kamailio package.

To fix this this add some modules (if they are not already selected)
that will kick of the build of the internal libraries in any case.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>