freeswitch-stable: move module selection to Build/Configure
Currently the routine that sets up the modules.conf file resides in
Build/Prepare.

When PKG_CONFIG_DEPENDS causes a reconfiguration, Build/Prepare is not
triggered. So if another module gets selected in menuconfig, this module
would not be added to modules.conf, meaning the rebuild would fail.

Add the routine to Build/Compile to address this.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>