pcapsipdump: revert uclibc++ commit
This reverts commit 6857bc21ebb7ab6408a968797f699921f0fb6441.

In hindsight I have to admit I did not correctly understand the
implications of the uclibc++.mk include.

The include allows a package to follow the user's choice regarding which
C++ library should be the standard. Linking against uClibc++ instead of
libstd++ is not a problem when running musl (which is what I had
incorrectly assumed), as both C++ libs are separate packages. And
uClibc++ is a lot smaller than libstd++, which is probably why it is
even the default C++ lib on OpenWrt currently.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>