asterisk-chan-lantiq: fix dep for nonshared builds
Despite all previous efforts chan-lantiq is not built by the build bots.
Checking the logs of a target build reveals that the chan-lantiq builds are
not even attempted.

During the target builds CONFIG_ALL is not set, meaning no regular package
gets selected. CONFIG_ALL_NONSHARED on the other hand is set. But
chan-lantiq still will not be selected, because it depends on "asterisk",
meaning asterisk needs to be selected first for chan-lantiq to become
available.

This commit changes the asterisk depend to "+asterisk", meaning asterisk
will be selected if chan-lantiq is selected.

This was tested with the SDK. A seed config file was added with the
following:

CONFIG_ALL_NONSHARED=y

Running menuselect showed that now chan-lantiq is selected
automatically. This should fix the problem.

The problem with "+asterisk" is that this breaks the asterisk menus for all
asterisk plugins/utilities that follow chan-lantiq alphabetically.  To
address this put the chan-lantiq modules into their own sub menu "Telephony
Lantiq". This is not nice, but the alternative would be to rename the
packages to something like asterisk13-z-chan-lantiq. This would make it
harder to find the packages. Maybe somebody has a better idea in the
future.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>