freeswitch-stable: include nls.mk, remove hack
Removes the uClibc iconv hack. Instead nls.mk is included to sort out
iconv depends.

mod_gsmopen isn't able to figure out if iconv's inbuf is const. In
OpenWrt this is the case with libiconv-full, not with libiconv-stub. To
aid mod_gsmopen update 270-fix-uclibc-iconv-in-gsmopen.patch so that we
can pass on if inbuf is const via CFLAGS.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>