freeswitch-stable: replace defaults with uci
This commit replaces /etc/default/freeswitch with /etc/config/freeswitch.
This way the init and hotplug configuration can be done with uci instead of
having to edit a file.

This also does away with the busybox ntpd warning. ntpd uses the same
configuration in system and it looks like busybox's ntpd is not used when
ntpd is installed.

Lastly some log strings are amended to start with a lowercase letter.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>