kamailio-5.x: amend mips2 tweak
Commit f84dda74e615c803e6a218cea83dd2a2bfaa67e5 ("kamailio-5.x: enable
FAST_LOCK for MIPS") turned out to be problematic, because it changed
the ARCH to "mips2" not only for "mips", but also for some "mipsel"
targets, which was unintentional.

Address this by filtering for "mips" specifically before setting the
variable.

Also, get rid of PKG_BUILD_PARALLEL, because adding it really didn't
change anything - due to the way "make" is called. Leave a comment to
prevent repetition (read: prevent _me_ from doing the same mistake again
in the future).

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>