freeswitch-stable: move hotplug into main pkg
There is no point in keeping the hotplug script in an extra package.
Move it to the main package.

The priority of the script is changed from 99 to 90 as the freeswitch
init script uses the same.

The postinstall script will check if 99-freeswitch exists. If it does a
warning is displayed with the suggestion to uninstall the hotplug
package.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>