asterisk-16.x-chan-lantiq: update snapshot
@arnysch pointed out on openwrt-devel that chan-lantiq fails to build
against kernel 4.19.x. He already fixed this upstream, so we can just
bump the snapshot to get chan-lantiq to compile again.

This also means that we can remove the malloc/free patch as it is
included upstream as well.

Finally, Build/Prepare and the file "default.exports" can be dropped as
the file got included upstream.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>