pcapsipdump: update to trunk snapshot
The last official release is quite old. But pcapsipdump's trunk has seen
a steady stream of fixes and improvements. This commit updates the
package to a current SVN snapshot.

MAKE_FLAGS is simplified. The deleted ARGS are provided by OpenWrt
buildroot, no need to specify them again. All that is needed is LIBS.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>