pcapsipdump: fix missing libbsd depend
When libbsd is available in staging it will be picked up by pcapsipdump
during the compile. The binary will link to libbsd, causing an
additional dependency. Currently the builds on the bots are failing
because of this.

The only function pcapsipdump is using from libbsd is strlcpy(). This
function is also provided by uClibc as well as musl.

Attached patch adds a detection mechanism that checks whether libc
provides strlcpy(). This way the dependency on libbsd is only required if
building against a libc which doesn't package strlcpy(), like glibc.
DEPENDS are updated to reflect that.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>