build: Add CircleCI
This adds CircleCI as Continous Integration. It will provide compile
test against ath79 target and together with it, it checks if the commit
starts with the package name and it checks if the Signed-off-by matches
the author of the commit.

For example, it can avoid mistakes with broken patches, mismatched changes.

This comes from packages feed and it is being used there and also in
LuCI repository.

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>