dahdi-linux: bump to version 3.1.0
This commit provides compatibility with kernel 5.4.

Except for the version bump this commit updates the HFC-S PCI driver to
not rely on the macro "dahdi_pci_module" anymore, because it has been
removed upstream in commit 4af6f69.

Also, a patch is added that makes the include of "<linux/pci-aspm.h>"
conditional on the kernel version, because starting with kernel 5.4 this
header has been removed and its contents merged into "pci.h".

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>