freeswitch: prevent libcheck use
Currently we disable some sofia tests to prevent build failure. An issue
was actually raised upstream ([1]) where this was discussed.

This cleared up some doubts. For instance it was established that
libcheck is used as a test framework, if the lib is detected. In OpenWrt
libcheck is available, so it may be in staging during the freeswitch
build.

Upstream also mentioned that libcheck is not always compatible with
freeswitch, which depends on the libcheck version.

This commit prevents libcheck detection. This is nicer than explicitly
disabling tests that fail.

[1] https://github.com/signalwire/freeswitch/issues/681

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>