spandsp3: new package
This commit adds an updated spandsp library. This is not a drop-in
replacement for the "old" spandsp library. Applications that want to use
it need to be updated a bit, according to upstream info [1].

In an effort to not cause any problems for OpenWrt, this package
installs the library and headers in the staging directory into
non-standard paths:

Library: $(STAGING_DIR/usr/lib/spandsp3/lib
Headers: $(STAGING_DIR/usr/lib/spandsp3/include

This way they are hidden aways from all normal users (like asterisk,
baresip etc.) and won't interfere.

To use the new spandsp library users can look for the spandsp3
pkg-config file (the regular spandsp installs spandsp.pc, this package
installs spandsp3.pc). This should be enough. The first such user will
likely be the freeswitch package, once it gets updated to 1.10.5.

[1] https://github.com/freeswitch/spandsp/issues/5

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>