siproxd: add max clients configuration and default
Add a configuration option SIPROXD_MAX_CLIENTS used for compile-time sizing
of the maximum supported proxy clients. Set this to a default value of 32,
which is more suitable for a small network or CPE router and uses much
less memory than the upstream value of 512 (e.g. RSS of 2 MB vs 17 MB).

Drop the following patch in favour of the new config option:

  * 005-reduce-rtpproxy-urlmap-size.patch

Also fix an incorrect configure option, and silence many warnings due to
unused VCS 'ident' variables in the sources.

Signed-off-by: Tony Ambardar <itugrok@yahoo.com>