baresip: use FILE_MODES instead of postinst
We can now set permissions directly thanks to OpenWrt having added
fakeroot support.. This also changes the permissions of /etc/baresip so
that baresip can write to this directory. For instance it wants to write
the file /etc/baresip/current_contact for which it needs the appropriate
permissions (baresip is started by user baresip).

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>