dahdi-tools: remove broken patch, disable unused
The patch "010-fix-non-glibc-builds.patch" is broken. The test for
HAVE_EXECINFO_H must happen after including "<autoconfig.h>".

Instead of fixing the patch this commit removes it. In lieu,
CONFIGURES_ARGS are added to the Makefile. These disable the building of
some tools that are not packaged anyway. Like this, "xpp/xtalk/debug.c"
is avoided, hence the patch becomes irrelevant.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>