asterisk: disable uriparser
Uriparser is a compliant URI parsing and handling library for RFC 3986.
For now, this is not included in OpenWrt feeds, but in Turris OS
uriparser is used as dependency for Updater-ng.

When Asterisk finds in build system there is uriparser or anything else mentioned in configure,
it tries to enable it by default. This applies to every package in
OpenWrt and because of that new packages which are added to OpenWrt tries to disable almost
everything by default. Because if someone adds library, some packages
are not compiled or tested.

Fixes:
Package asterisk is missing dependencies for the following libraries:
liburiparser.so.1

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>