kamailio: prevent arm -march overrides
src/Makefile.defs adds "-march=..." to target flags for ARM targets.
This can clash with the OpenWrt target flags in
CONFIG_TARGET_OPTIMIZATION. If for example the latter is set to
"-mcpu=cortex-a9" the following warning is issued:

cc1: warning: switch '-mcpu=cortex-a9' conflicts with '-march=armv5t' switch

This commit removes these flags. The addition is unwanted when
cross-compiling.

An issue was raised for this recently, see [1].

[1] https://github.com/openwrt/packages/issues/16135

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>