freeswitch: don't start with explicit group
Currently the freeswitch init script starts the service with the group
"freeswitch". Like this, even if the freeswitch user is part of other
groups, freeswitch will not be able to make use of them. So for instance
if you add the user to the group "dialout", freeswitch will run under
group "freeswitch", instead of "freeswitch" _and_ "dialout".

Not specifying the group gets rid of this limitation.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>