asterisk: don't start with explicit group
Currently the asterisk init script starts the service with the group
"asterisk". Like this, even if the asterisk user is part of other groups,
asterisk will not be able to make use of them. So for instance if you add the
user to the group "dialout", asterisk will run under group "asterisk", instead
of "asterisk" _and_ "dialout".

Not specifying the group gets rid of this limitation.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>