coturn: improve reproducibility (Ubuntu build host fix)
coturn package is reproducible unless the build OS is Ubuntu

coturn configure script detects Ubuntu build host([1]) and changes
compilation flags so produced OpenWrt binaries are different on
Ubuntu and any other build OS (e.g. Debian). It might be necessary
for native build but this check is not valid for cross-compiling.

This patch set LIBEV_OK=1 to generate the same binaries on Ubuntu
and other build OS

Refs:
[1] https://github.com/coturn/coturn/blob/upstream/4.5.2/configure#L426-L435

Signed-off-by: Sergey V. Lobanov <sergey@lobanov.in>