baresip: only build gst with AUDIO_SUPPORT enabled
Most gstreamer packages aren't build when the target does not have
AUDIO_SUPPORT. This can be seen with realtek target (4kec).

This commit makes the gst plugins depend on AUDIO_SUPPORT, because
otherwise the build dies with this error:

Package baresip-mod-gst-video is missing dependencies for the following libraries:
libgstapp-1.0.so.0

Note: compiling baresip on a target without AUDIO_SUPPORT may be
questionable anyway.

Close #730

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>