re: Bump to version 3.16.0
The previous version 2.0.1 is about 7 years old now and a lot has
changed. Various fixes, improvements and features have been added.

The build system has moved to cmake and pkg-config is used,
which makes it a lot easier to build from the openwrt buildroot.
Therefore a lot of the extra code is removed from it's Makefile

Librem is now integrated into libre.
No more patches are required.

Signed-off-by: Daniel Danzberger <dd@embedd.com>