baresip: Bump to version 3.16.0
The previous version 1.1.0 is about 7 years old now and a lot has
changed. Various fixes, improvements and features have been added.

The build system has moved to cmake and pkg-config is used,
which makes it a lot easier to build from the openwrt buildroot.
Therefore a lot of the extra code is removed from it's Makefile.

The librem dependency has been removed, since newer versions of libre
contain all the librem features.

The parallel build patch is no longer required.

Signed-off-by: Daniel Danzberger <dd@embedd.com>