spandsp3: fix leaking host header for build tools
It was discovered that spandsp3 leaks host header on building build
tools.

This was caused by the Makefile not permitting to pass custom header on
compiling build tools. To be more precise it was possible to leak host
header for the tiff library for the tiffio.h header.

Add pending patch to fix this and not depend on the host system header.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>